requirements.txt
asgiref==3.5.1
Django==3.2.7
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Pillow==9.1.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-decouple==3.6
pytz==2022.1
Serializer==0.2.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
